enable gpg validation :secret:

+++ .config/ranger/plugins/decrypt.py
@@ -13,21 +13,27 @@ class decrypt(Command):
     """
 
     def execute(self):
-        gpg = GPG(gnupghome=os.path.join(os.path.expanduser('~'), '.gnupg'))
+        gpg = GPG(gnupghome=os.path.join(os.path.expanduser("~"), ".gnupg"))
 
         paths = [os.path.basename(f.path) for f in self.fm.thistab.get_selection()]
 
-        for p in [p for p in paths if p.endswith('gpg')]:
-            with open(p, 'rb') as enc:
+        for p in [p for p in paths if p.endswith("gpg")]:
+            dec_b = None
+
+            with open(p, "rb") as enc:
                 dec_b = gpg.decrypt_file(enc)
 
+            if not dec_b.ok:
+                self.fm.notify("GPG Error: " + dec_b.status)
+                return
+
             out_fname = os.path.splitext(p)[0]
-            with open(out_fname, 'wb+') as dec_f:
+            with open(out_fname, "wb+") as dec_f:
                 dec_f.write(dec_b.data)
 
-            if self.arg(1) != 'true':
+            if self.arg(1) != "true":
                 os.remove(p)
 
             if tarfile.is_tarfile(out_fname):
-                tarfile.open(out_fname).extractall(path='.')
+                tarfile.open(out_fname).extractall(path=".")
                 os.remove(out_fname)
